// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PaymentAttemptStatus = {
  PENDING: 'PENDING',
  AUTHORIZED: 'AUTHORIZED',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED'
} as const

export type PaymentAttemptStatus = (typeof PaymentAttemptStatus)[keyof typeof PaymentAttemptStatus]


export const RefundStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type RefundStatus = (typeof RefundStatus)[keyof typeof RefundStatus]


export const OrderStatusKey = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  CANCELED: 'CANCELED'
} as const

export type OrderStatusKey = (typeof OrderStatusKey)[keyof typeof OrderStatusKey]


export const CartEventType = {
  ADD_TO_CART: 'ADD_TO_CART',
  UPDATE_QUANTITY: 'UPDATE_QUANTITY',
  REMOVE_FROM_CART: 'REMOVE_FROM_CART',
  CLEAR_CART: 'CLEAR_CART',
  LOCK_CART: 'LOCK_CART',
  UNLOCK_CART: 'UNLOCK_CART'
} as const

export type CartEventType = (typeof CartEventType)[keyof typeof CartEventType]


export const RatingScore = {
  ONE: 'ONE',
  TWO: 'TWO',
  THREE: 'THREE',
  FOUR: 'FOUR',
  FIVE: 'FIVE'
} as const

export type RatingScore = (typeof RatingScore)[keyof typeof RatingScore]


export const TokenType = {
  REFRESH: 'REFRESH',
  VERIFICATION: 'VERIFICATION',
  FORGOT_PASSWORD: 'FORGOT_PASSWORD'
} as const

export type TokenType = (typeof TokenType)[keyof typeof TokenType]


export const RoleKey = {
  ADMIN: 'ADMIN',
  CUSTOMER: 'CUSTOMER',
  RESTAURANT_MANAGER: 'RESTAURANT_MANAGER'
} as const

export type RoleKey = (typeof RoleKey)[keyof typeof RoleKey]